+++ estampala.server.core/catalog/src/main/java/catalog/services/ShirtMaterialService.java
@@ -35,18 +35,31 @@ public class ShirtMaterialService {
 	}
 	
 	public ShirtMaterial save(ShirtMaterial item) {
+		if (item != null){
+			item.setId(UUID.randomUUID());
+		}
 		return repository.save(item);
 	}	
 	
 	public ShirtMaterial update(ShirtMaterial item) {
-		return repository.save(item);
+		if (item != null){
+			return repository.save(item);
+		}
+		
+		return null;
 	}	
 	
-	public void delete(UUID id){		
-		repository.delete(id);						
+	public void delete(UUID id){
+		if(id != null){
+			repository.delete(id);
+		}
 	}
 	
-	public boolean exists(UUID id){		
-		return repository.exists(id);						
+	public boolean exists(UUID id){
+		if(id != null){
+			return repository.exists(id);
+		}
+		
+		return false;
 	}
 }
